pyproject.toml
[tool.poetry]
name = "tobi-conversational-sales"
version = "0.1.0"
description = "Conversational Sales Bot for Vodafone UK"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
python-multipart = "^0.0.6"
openai = "^1.3.0"
playwright = "^1.40.0"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
numpy = "^2.1.0"
tiktoken = "^0.9.0"
langchain = "^0.3.26"
langchain-community = "^0.3.27"
langchain-openai = "^0.3.28"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.10.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 